refactor(cli): (Request) => Response runtime core; node listener becomes an adapter
